requirement.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.0
cffi==1.15.1
channels==3.0.0
channels-redis==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.4
daphne==3.0.2
Django==3.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.4
mysqlclient==2.1.1
numpy==1.23.5
pandas==1.5.2
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing-extensions==4.4.0
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.5.2
